chore: improve some nits
