+++ crates/curve/src/tests/curve/bn254.rs
@@ -1,9 +1,10 @@
-use crate::{tests::mimc::test_mimc_groth_16, CurveBasicOperations};
 use ark_bn254::{Bn254, Fr};
 use ark_ff::{Field, FromBytes, ToBytes};
 use ark_groth16::{verify_proof, PreparedVerifyingKey, Proof, VerifyingKey};
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
 use ark_std::{ops::MulAssign, test_rng, vec::Vec, UniformRand};
+
+use crate::{tests::mimc::test_mimc_groth_16, CurveBasicOperations};
 
 /// BN254 ADD
 pub fn bn254_add() {
