+++ crates/curve/src/tests/curve/mod.rs
@@ -1,4 +1,5 @@
 //! Tests as exports
+
 mod bls12_377;
 mod bls12_381;
 mod bn254;
