+++ crates/eddsa/src/lib.rs
@@ -1,13 +1,13 @@
 #![no_std]
+
 extern crate alloc;
 
-#[allow(dead_code)]
 mod etec;
 mod jubjub;
 mod verifier;
 mod w_naf;
 
-pub use etec::*;
-pub use jubjub::*;
-pub use verifier::*;
-pub use w_naf::*;
+pub use self::etec::*;
+pub use self::jubjub::*;
+pub use self::verifier::*;
+pub use self::w_naf::*;
