+++ crates/merkle-tree/src/mimc.rs
@@ -1,4 +1,5 @@
 use alloc::vec::Vec;
+
 use once_cell::sync::Lazy;
 use tiny_keccak::{Hasher, Keccak};
 use zkp_u256::{Zero, U256};
